Revert "feat: add chat tab to bottom navigation (#3114)"
This reverts commit 5ce2a6a325065d300d9f0d5c36a3452d459651bf.
