+++ app/lib/pages/apps/app_detail/app_detail.dart
@@ -15,7 +15,6 @@ import 'package:omi/pages/apps/markdown_viewer.dart';
 import 'package:omi/pages/chat/page.dart';
 import 'package:omi/pages/apps/providers/add_app_provider.dart';
 import 'package:omi/providers/app_provider.dart';
-import 'package:omi/providers/home_provider.dart';
 import 'package:omi/providers/message_provider.dart';
 import 'package:omi/utils/analytics/mixpanel.dart';
 import 'package:omi/utils/other/temp.dart';
@@ -309,10 +308,14 @@ class _AppDetailPageState extends State<AppDetailPage> {
                           messageProvider.sendInitialAppMessage(selectedApp);
                         }
 
-                        // Navigate to chat tab
+                        // Navigate directly to chat page
                         if (mounted) {
-                          Navigator.pop(context);
-                          context.read<HomeProvider>().setIndex(2);
+                          Navigator.push(
+                            context,
+                            MaterialPageRoute(
+                              builder: (context) => const ChatPage(isPivotBottom: false),
+                            ),
+                          );
                         }
                       } finally {
                         if (mounted) {
