+++ app/lib/pages/chat/page.dart
@@ -1,21 +1,27 @@
+import 'package:cached_network_image/cached_network_image.dart';
+import 'package:collection/collection.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/rendering.dart';
 import 'package:flutter/scheduler.dart';
 import 'package:flutter/services.dart';
 import 'package:font_awesome_flutter/font_awesome_flutter.dart';
+import 'package:pull_down_button/pull_down_button.dart';
 import 'package:omi/backend/http/api/messages.dart';
 import 'package:omi/backend/preferences.dart';
 import 'package:omi/backend/schema/app.dart';
 import 'package:omi/backend/schema/conversation.dart';
 import 'package:omi/backend/schema/message.dart';
+import 'package:omi/gen/assets.gen.dart';
 import 'package:omi/pages/chat/select_text_screen.dart';
 import 'package:omi/pages/chat/widgets/ai_message.dart';
 import 'package:omi/pages/chat/widgets/user_message.dart';
 import 'package:omi/pages/chat/widgets/voice_recorder_widget.dart';
+import 'package:omi/pages/home/page.dart';
 import 'package:omi/providers/connectivity_provider.dart';
 import 'package:omi/providers/home_provider.dart';
 import 'package:omi/providers/conversation_provider.dart';
 import 'package:omi/providers/message_provider.dart';
+import 'package:omi/providers/app_provider.dart';
 import 'package:omi/utils/alerts/app_snackbar.dart';
 import 'package:omi/utils/analytics/mixpanel.dart';
 import 'package:omi/utils/other/temp.dart';
@@ -52,6 +58,7 @@ class ChatPageState extends State<ChatPage> with AutomaticKeepAliveClientMixin, 
   late List<App> apps;
 
   final scaffoldKey = GlobalKey<ScaffoldState>();
+  final GlobalKey _appButtonKey = GlobalKey();
 
   @override
   bool get wantKeepAlive => true;
@@ -60,7 +67,7 @@ class ChatPageState extends State<ChatPage> with AutomaticKeepAliveClientMixin, 
   void initState() {
     apps = prefs.appsList;
     scrollController = ScrollController();
-    textFieldFocusNode = context.read<HomeProvider>().chatFieldFocusNode;
+    textFieldFocusNode = FocusNode();
     textController.addListener(() {
       setState(() {});
     });
@@ -112,6 +119,7 @@ class ChatPageState extends State<ChatPage> with AutomaticKeepAliveClientMixin, 
   void dispose() {
     textController.dispose();
     scrollController.dispose();
+    textFieldFocusNode.dispose();
     super.dispose();
   }
 
@@ -124,19 +132,18 @@ class ChatPageState extends State<ChatPage> with AutomaticKeepAliveClientMixin, 
         return Scaffold(
           key: scaffoldKey,
           backgroundColor: Theme.of(context).colorScheme.primary,
+          appBar: _buildAppBar(context, provider),
           // endDrawer: _buildSessionsDrawer(context),
           body: GestureDetector(
             onTap: () {
               // Hide keyboard when tapping outside textfield
               FocusScope.of(context).unfocus();
             },
-            child: Stack(
+            child: Column(
               children: [
-                Column(
-                  children: [
-                    // Messages area - takes up remaining space
-                    Expanded(
-                      child: provider.isLoadingMessages && !provider.hasCachedMessages
+                // Messages area - takes up remaining space
+                Expanded(
+                  child: provider.isLoadingMessages && !provider.hasCachedMessages
                       ? Column(
                           mainAxisAlignment: MainAxisAlignment.center,
                           children: [
@@ -442,11 +449,7 @@ class ChatPageState extends State<ChatPage> with AutomaticKeepAliveClientMixin, 
                             left: 0,
                             right: 16,
                             top: provider.selectedFiles.isNotEmpty ? 0 : 16,
-                            bottom: widget.isPivotBottom 
-                                ? 20 
-                                : textFieldFocusNode.hasFocus 
-                                    ? MediaQuery.of(context).viewInsets.bottom + 20
-                                    : MediaQuery.of(context).padding.bottom + 80,
+                            bottom: widget.isPivotBottom ? 20 : (textFieldFocusNode.hasFocus ? 20 : 40),
                           ),
                           child: Row(
                             crossAxisAlignment: CrossAxisAlignment.end,
@@ -621,32 +624,9 @@ class ChatPageState extends State<ChatPage> with AutomaticKeepAliveClientMixin, 
                 ),
               ],
             ),
-            // Loading messages indicator
-            if (provider.isLoadingMessages)
-              Positioned(
-                top: MediaQuery.of(context).padding.top + 10,
-                left: 0,
-                right: 0,
-                child: Container(
-                  margin: const EdgeInsets.symmetric(horizontal: 20),
-                  padding: const EdgeInsets.symmetric(vertical: 8, horizontal: 16),
-                  decoration: BoxDecoration(
-                    color: Colors.green,
-                    borderRadius: BorderRadius.circular(20),
-                  ),
-                  child: const Center(
-                    child: Text(
-                      'Syncing messages with server...',
-                      style: TextStyle(color: Colors.white, fontSize: 12),
-                    ),
-                  ),
-                ),
-              ),
-          ],
-        ),
-          )
-      );
-    },
+          ),
+        );
+      },
     );
   }
 
@@ -693,6 +673,354 @@ class ChatPageState extends State<ChatPage> with AutomaticKeepAliveClientMixin, 
   }
 
   scrollToBottom() => _moveListToBottom();
+
+  void _handleAppSelection(String? val, AppProvider provider) {
+    if (val == null || val == provider.selectedChatAppId) {
+      return;
+    }
+
+    // Unfocus the text field to prevent keyboard issues
+    textFieldFocusNode.unfocus();
+
+    // clear chat
+    if (val == 'clear_chat') {
+      _showClearChatDialog();
+      return;
+    }
+
+    // enable apps - navigate back to home and show apps page
+    if (val == 'enable') {
+      _navigateToAppsPage();
+      return;
+    }
+
+    // select app by id
+    _selectApp(val, provider);
+  }
+
+  void _showClearChatDialog() {
+    if (!mounted) return;
+
+    showDialog(
+      context: context,
+      builder: (ctx) {
+        return getDialog(context, () {
+          Navigator.of(context).pop();
+        }, () {
+          if (mounted) {
+            context.read<MessageProvider>().clearChat();
+            Navigator.of(context).pop();
+          }
+        }, "Clear Chat?", "Are you sure you want to clear the chat? This action cannot be undone.");
+      },
+    );
+  }
+
+  void _navigateToAppsPage() {
+    if (!mounted) return;
+
+    MixpanelManager().pageOpened('Chat Apps');
+    Navigator.of(context).pushReplacement(
+      MaterialPageRoute(
+        builder: (context) => const HomePageWrapper(navigateToRoute: '/apps'),
+      ),
+    );
+  }
+
+  void _selectApp(String appId, AppProvider appProvider) async {
+    if (!mounted) return;
+
+    // Mark that we're no longer on initial load to prevent auto-focus
+    _isInitialLoad = false;
+
+    // Store references before async operation
+    final messageProvider = mounted ? context.read<MessageProvider>() : null;
+    if (messageProvider == null) return;
+
+    // Set the selected app
+    appProvider.setSelectedChatAppId(appId);
+
+    // Add a small delay to let the keyboard animation complete
+    // This prevents the widget from being unmounted during the keyboard transition
+    await Future.delayed(const Duration(milliseconds: 100));
+
+    // Check if widget is still mounted after delay
+    if (!mounted) return;
+
+    // Perform async operation
+    await messageProvider.refreshMessages(dropdownSelected: true);
+
+    // Check if widget is still mounted before proceeding
+    if (!mounted) return;
+
+    // Get the selected app and send initial message if needed
+    var app = appProvider.getSelectedApp();
+    if (messageProvider.messages.isEmpty) {
+      messageProvider.sendInitialAppMessage(app);
+    }
+  }
+
+  PreferredSizeWidget _buildAppBar(BuildContext context, MessageProvider provider) {
+    return AppBar(
+      backgroundColor: Theme.of(context).colorScheme.surface,
+      elevation: 0,
+      leading: IconButton(
+        icon: const Icon(Icons.arrow_back, color: Colors.white),
+        onPressed: () => Navigator.of(context).pop(),
+      ),
+      title: Consumer<AppProvider>(
+        builder: (context, appProvider, child) {
+          return _buildAppSelection(context, appProvider);
+        },
+      ),
+      centerTitle: true,
+      actions: const [
+        // IconButton(
+        //   icon: const Icon(Icons.history, color: Colors.white),
+        //   onPressed: () {
+        //     HapticFeedback.mediumImpact();
+        //     // Dismiss keyboard before opening drawer
+        //     FocusScope.of(context).unfocus();
+        //     scaffoldKey.currentState?.openEndDrawer();
+        //   },
+        // ),
+      ],
+      bottom: provider.isLoadingMessages
+          ? PreferredSize(
+              preferredSize: const Size.fromHeight(32),
+              child: Container(
+                width: double.infinity,
+                height: 32,
+                color: Colors.green,
+                child: const Center(
+                  child: Text(
+                    'Syncing messages with server...',
+                    style: TextStyle(color: Colors.white, fontSize: 12),
+                  ),
+                ),
+              ),
+            )
+          : null,
+    );
+  }
+
+  void _showAppsMenu(BuildContext ctx, AppProvider appProvider) {
+    final renderBox = _appButtonKey.currentContext?.findRenderObject() as RenderBox?;
+    if (renderBox == null) return;
+
+    final overlay = Overlay.of(ctx, rootOverlay: true);
+
+    final buttonOffset = renderBox.localToGlobal(Offset.zero);
+    final buttonSize = renderBox.size;
+    final screenSize = MediaQuery.of(ctx).size;
+
+    const double menuWidth = 260;
+    const double maxMenuHeight = 250;
+
+    double desiredTop = buttonOffset.dy + buttonSize.height + 8;
+    if ((desiredTop + maxMenuHeight) > screenSize.height) {
+      desiredTop = buttonOffset.dy - maxMenuHeight - 8;
+      if (desiredTop < 0) desiredTop = 8;
+    }
+
+    late OverlayEntry entry;
+
+    final controller = AnimationController(
+      vsync: this,
+      duration: const Duration(milliseconds: 200),
+    );
+    final curved = CurvedAnimation(parent: controller, curve: Curves.easeOut);
+
+    entry = OverlayEntry(
+      builder: (context) {
+        return Consumer<MessageProvider>(
+          builder: (context, msgProvider, _) {
+            return Stack(
+              children: [
+                Positioned.fill(
+                  child: GestureDetector(
+                    behavior: HitTestBehavior.translucent,
+                    onTap: () {
+                      controller.reverse().then((_) => entry.remove());
+                    },
+                    child: Container(color: Colors.transparent),
+                  ),
+                ),
+                Positioned(
+                  left: buttonOffset.dx + (buttonSize.width - menuWidth) / 2,
+                  top: desiredTop,
+                  child: AnimatedBuilder(
+                    animation: curved,
+                    builder: (context, child) {
+                      return Transform.scale(
+                        scale: curved.value,
+                        alignment: Alignment.topCenter,
+                        child: Opacity(
+                          opacity: curved.value,
+                          child: child,
+                        ),
+                      );
+                    },
+                    child: Material(
+                      color: Colors.grey[900],
+                      borderRadius: BorderRadius.circular(12),
+                      elevation: 8,
+                      child: SizedBox(
+                        width: menuWidth,
+                        height: maxMenuHeight,
+                        child: PullDownMenu(
+                          items: [
+                            PullDownMenuItem(
+                              title: 'Clear Chat',
+                              iconWidget: const Icon(Icons.delete, color: Colors.redAccent, size: 16),
+                              onTap: () {
+                                controller.reverse().then((_) {
+                                  entry.remove();
+                                  _handleAppSelection('clear_chat', appProvider);
+                                });
+                              },
+                            ),
+                            PullDownMenuItem(
+                              title: 'Enable Apps',
+                              iconWidget: const Icon(Icons.arrow_forward_ios, color: Colors.white60, size: 16),
+                              onTap: () {
+                                controller.reverse().then((_) {
+                                  entry.remove();
+                                  _handleAppSelection('enable', appProvider);
+                                });
+                              },
+                            ),
+                            PullDownMenuItem(
+                              title: 'Omi',
+                              iconWidget: _getOmiAvatar(),
+                              onTap: () {
+                                controller.reverse().then((_) {
+                                  entry.remove();
+                                  _handleAppSelection('no_selected', appProvider);
+                                });
+                              },
+                              subtitle:
+                                  msgProvider.chatApps.firstWhereOrNull((a) => a.id == appProvider.selectedChatAppId) ==
+                                          null
+                                      ? 'Selected'
+                                      : null,
+                            ),
+                            ...msgProvider.chatApps.map(
+                              (app) => PullDownMenuItem(
+                                title: app.getName(),
+                                iconWidget: _getAppAvatar(app),
+                                onTap: () {
+                                  controller.reverse().then((_) {
+                                    entry.remove();
+                                    _handleAppSelection(app.id, appProvider);
+                                  });
+                                },
+                                subtitle: appProvider.selectedChatAppId == app.id ? 'Selected' : null,
+                              ),
+                            )
+                          ],
+                        ),
+                      ),
+                    ),
+                  ),
+                ),
+              ],
+            );
+          },
+        );
+      },
+    );
+
+    overlay.insert(entry);
+    controller.forward();
+  }
+
+  Widget _buildAppSelection(BuildContext context, AppProvider provider) {
+    final messageProvider = Provider.of<MessageProvider>(context, listen: false);
+    var selectedApp = messageProvider.chatApps.firstWhereOrNull((app) => app.id == provider.selectedChatAppId);
+
+    return GestureDetector(
+      key: _appButtonKey,
+      onTap: () {
+        HapticFeedback.mediumImpact();
+        _showAppsMenu(context, provider);
+      },
+      child: Row(
+        crossAxisAlignment: CrossAxisAlignment.center,
+        mainAxisAlignment: MainAxisAlignment.center,
+        mainAxisSize: MainAxisSize.min,
+        children: [
+          selectedApp != null ? _getAppAvatar(selectedApp) : _getOmiAvatar(),
+          const SizedBox(width: 8),
+          Container(
+            constraints: const BoxConstraints(maxWidth: 100),
+            child: Text(
+              selectedApp != null ? selectedApp.getName() : "Omi",
+              style: const TextStyle(color: Colors.white, fontSize: 16),
+              overflow: TextOverflow.fade,
+            ),
+          ),
+          const SizedBox(width: 8),
+          const SizedBox(
+            width: 16,
+            child: Icon(Icons.keyboard_arrow_down, color: Colors.white60, size: 16),
+          ),
+        ],
+      ),
+    );
+  }
+
+  Widget _getAppAvatar(App app) {
+    return CachedNetworkImage(
+      imageUrl: app.getImageUrl(),
+      imageBuilder: (context, imageProvider) {
+        return CircleAvatar(
+          backgroundColor: Colors.white,
+          radius: 12,
+          backgroundImage: imageProvider,
+        );
+      },
+      errorWidget: (context, url, error) {
+        return const CircleAvatar(
+          backgroundColor: Colors.white,
+          radius: 12,
+          child: Icon(Icons.error_outline_rounded),
+        );
+      },
+      progressIndicatorBuilder: (context, url, progress) => CircleAvatar(
+        backgroundColor: Colors.white,
+        radius: 12,
+        child: CircularProgressIndicator(
+          value: progress.progress,
+          valueColor: const AlwaysStoppedAnimation<Color>(Colors.white),
+        ),
+      ),
+    );
+  }
+
+  Widget _getOmiAvatar() {
+    return Container(
+      decoration: BoxDecoration(
+        image: DecorationImage(
+          image: AssetImage(Assets.images.background.path),
+          fit: BoxFit.cover,
+        ),
+        borderRadius: const BorderRadius.all(Radius.circular(16.0)),
+      ),
+      height: 24,
+      width: 24,
+      child: Stack(
+        alignment: Alignment.center,
+        children: [
+          Image.asset(
+            Assets.images.herologo.path,
+            height: 16,
+            width: 16,
+          ),
+        ],
+      ),
+    );
+  }
 
   void _showIOSStyleActionSheet(BuildContext context) {
     showModalBottomSheet(
