+++ app/lib/pages/home/widgets/battery_info_widget.dart
@@ -30,11 +30,9 @@ class BatteryInfoWidget extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return Selector<HomeProvider, int>(
-      selector: (context, state) => state.selectedIndex,
-      builder: (context, selectedIndex, child) {
-        final isConversationPage = selectedIndex == 0;
-        final isChatPage = selectedIndex == 2;
+    return Selector<HomeProvider, bool>(
+      selector: (context, state) => state.selectedIndex == 0,
+      builder: (context, isMemoriesPage, child) {
         return Consumer<DeviceProvider>(
           builder: (context, deviceProvider, child) {
             if (deviceProvider.connectedDevice != null) {
@@ -126,11 +124,10 @@ class BatteryInfoWidget extends StatelessWidget {
                         ),
                       ),
                       const SizedBox(width: 8.0),
-                      if (!isChatPage)
-                        Text(
-                          "Disconnected",
-                          style: Theme.of(context).textTheme.bodyMedium!.copyWith(color: Colors.white70),
-                        ),
+                      Text(
+                        "Disconnected",
+                        style: Theme.of(context).textTheme.bodyMedium!.copyWith(color: Colors.white70),
+                      ),
                     ],
                   ),
                 ),
@@ -158,13 +155,13 @@ class BatteryInfoWidget extends StatelessWidget {
                         width: MediaQuery.sizeOf(context).width * 0.05,
                         height: MediaQuery.sizeOf(context).width * 0.05,
                       ),
-                      isConversationPage ? const SizedBox(width: 8) : const SizedBox.shrink(),
-                      deviceProvider.isConnecting && isConversationPage
+                      isMemoriesPage ? const SizedBox(width: 8) : const SizedBox.shrink(),
+                      deviceProvider.isConnecting && isMemoriesPage
                           ? Text(
                               "Searching",
                               style: Theme.of(context).textTheme.bodyMedium!.copyWith(color: Colors.white),
                             )
-                          : isConversationPage
+                          : isMemoriesPage
                               ? Text(
                                   "Connect Device",
                                   style: Theme.of(context).textTheme.bodyMedium!.copyWith(color: Colors.white),
