+++ app/lib/pages/memories/page.dart
@@ -57,7 +57,6 @@ class MemoriesPageState extends State<MemoriesPage> with AutomaticKeepAliveClien
   bool get wantKeepAlive => true;
 
   final TextEditingController _searchController = TextEditingController();
-  final FocusNode _searchFocusNode = FocusNode();
   MemoryCategory? _selectedCategory;
   final ScrollController _scrollController = ScrollController();
 
@@ -70,7 +69,6 @@ class MemoriesPageState extends State<MemoriesPage> with AutomaticKeepAliveClien
   @override
   void dispose() {
     _searchController.dispose();
-    _searchFocusNode.dispose();
     _scrollController.dispose();
     _removeDeleteNotification();
     super.dispose();
@@ -238,10 +236,6 @@ class MemoriesPageState extends State<MemoriesPage> with AutomaticKeepAliveClien
         return PopScope(
           canPop: true,
           child: Scaffold(
-            appBar: AppBar(
-              title: const Text('Memories'),
-              backgroundColor: Theme.of(context).colorScheme.primary,
-            ),
             backgroundColor: Theme.of(context).colorScheme.primary,
             body: RefreshIndicator(
               onRefresh: () async {
@@ -339,7 +333,7 @@ class MemoriesPageState extends State<MemoriesPage> with AutomaticKeepAliveClien
                                         padding: WidgetStateProperty.all(
                                           const EdgeInsets.symmetric(horizontal: 12, vertical: 4),
                                         ),
-                                        focusNode: _searchFocusNode,
+                                        focusNode: home.memoriesSearchFieldFocusNode,
                                         controller: _searchController,
                                         trailing: provider.searchQuery.isNotEmpty
                                             ? [
@@ -579,7 +573,7 @@ class MemoriesPageState extends State<MemoriesPage> with AutomaticKeepAliveClien
                           )
                         else
                           SliverPadding(
-                            padding: const EdgeInsets.only(top: 8, left: 16, right: 16, bottom: 80),
+                            padding: const EdgeInsets.only(top: 8, left: 16, right: 16, bottom: 120),
                             sliver: SliverList(
                               delegate: SliverChildBuilderDelegate(
                                 (context, index) {
@@ -622,7 +616,7 @@ class MemoriesPageState extends State<MemoriesPage> with AutomaticKeepAliveClien
 
   Widget _buildShimmerMemoryList() {
     return Padding(
-      padding: const EdgeInsets.only(top: 8, left: 16, right: 16, bottom: 80),
+      padding: const EdgeInsets.only(top: 8, left: 16, right: 16, bottom: 120),
       child: ListView.builder(
         itemCount: 8, // Show 8 shimmer items
         itemBuilder: (context, index) {
