+++ app/lib/providers/home_provider.dart
@@ -12,9 +12,11 @@ class HomeProvider extends ChangeNotifier {
   final FocusNode chatFieldFocusNode = FocusNode();
   final FocusNode appsSearchFieldFocusNode = FocusNode();
   final FocusNode convoSearchFieldFocusNode = FocusNode();
+  final FocusNode memoriesSearchFieldFocusNode = FocusNode();
   bool isAppsSearchFieldFocused = false;
   bool isChatFieldFocused = false;
   bool isConvoSearchFieldFocused = false;
+  bool isMemoriesSearchFieldFocused = false;
   bool hasSpeakerProfile = true;
   bool isLoading = false;
   String userPrimaryLanguage = SharedPreferencesUtil().userPrimaryLanguage;
@@ -67,28 +69,20 @@ class HomeProvider extends ChangeNotifier {
     chatFieldFocusNode.addListener(_onFocusChange);
     appsSearchFieldFocusNode.addListener(_onFocusChange);
     convoSearchFieldFocusNode.addListener(_onFocusChange);
+    memoriesSearchFieldFocusNode.addListener(_onFocusChange);
   }
 
   void _onFocusChange() {
     isChatFieldFocused = chatFieldFocusNode.hasFocus;
     isAppsSearchFieldFocused = appsSearchFieldFocusNode.hasFocus;
     isConvoSearchFieldFocused = convoSearchFieldFocusNode.hasFocus;
+    isMemoriesSearchFieldFocused = memoriesSearchFieldFocusNode.hasFocus;
     notifyListeners();
   }
 
   void setIndex(int index) {
     selectedIndex = index;
     notifyListeners();
-
-    if (index == 2) {
-      WidgetsBinding.instance.addPostFrameCallback((_) {
-        Future.delayed(const Duration(milliseconds: 100), () {
-          if (chatFieldFocusNode.canRequestFocus && !chatFieldFocusNode.hasFocus) {
-            chatFieldFocusNode.requestFocus();
-          }
-        });
-      });
-    }
   }
 
   void setIsLoading(bool loading) {
@@ -187,6 +181,8 @@ class HomeProvider extends ChangeNotifier {
     chatFieldFocusNode.removeListener(_onFocusChange);
     appsSearchFieldFocusNode.removeListener(_onFocusChange);
     convoSearchFieldFocusNode.removeListener(_onFocusChange);
+    memoriesSearchFieldFocusNode.removeListener(_onFocusChange);
+    memoriesSearchFieldFocusNode.dispose();
     chatFieldFocusNode.dispose();
     appsSearchFieldFocusNode.dispose();
     convoSearchFieldFocusNode.dispose();
